Move printing code to a separate file
