+++ riscv_analysis_cli/src/main.rs
@@ -1,3 +1,6 @@
+mod printer;
+use printer::*;
+
 use std::fmt::Display;
 use std::io::Write;
 use std::vec;
@@ -8,7 +11,7 @@ use bat::{Input, PrettyPrinter};
 use colored::Colorize;
 use riscv_analysis::fix::{fix_stack, Manipulation};
 use riscv_analysis::parser::{Info, LabelString, RVParser, With};
-use riscv_analysis::passes::{DiagnosticItem, SeverityLevel};
+use riscv_analysis::passes::DiagnosticItem;
 use std::path::PathBuf;
 use uuid::Uuid;
 
@@ -267,59 +270,6 @@ impl FileReader for IOFileReader {
     }
 }
 
-trait ErrorDisplay {
-    fn display_errors<T: FileReader + Clone>(&self, parser: &RVParser<T>);
-}
-
-impl ErrorDisplay for Vec<DiagnosticItem> {
-    fn display_errors<T: FileReader + Clone>(&self, parser: &RVParser<T>) {
-        for err in self {
-            let filename = parser
-                .reader
-                .get_filename(err.file)
-                .unwrap_or("unknown".to_owned());
-
-            let level = match err.level {
-                SeverityLevel::Error => "ERROR",
-                SeverityLevel::Warning => "WARNING",
-                SeverityLevel::Information => "INFO",
-                SeverityLevel::Hint => "HINT",
-            };
-
-            if let Some(text) = parser.reader.get_text(err.file) {
-                PrettyPrinter::new()
-                    .input(
-                        Input::from_reader(text.as_bytes())
-                            .kind(format!("{} in file", level))
-                            .name(filename.clone()),
-                    )
-                    .header(true)
-                    .line_numbers(true)
-                    .grid(true)
-                    .paging_mode(bat::PagingMode::Never)
-                    .line_ranges(LineRanges::from(vec![LineRange::new(
-                        err.range.start.line + 1,
-                        err.range.start.line + 1,
-                    )]))
-                    .print()
-                    .unwrap();
-
-                // print range arrows
-                println!(
-                    "       {}",
-                    " ".repeat(err.range.start.column)
-                        + &"^".repeat(err.range.end.column - err.range.start.column)
-                );
-            }
-
-            print!(
-                "       {}\n       {}\n",
-                err.title.bold(),
-                err.long_description
-            );
-        }
-    }
-}
 fn main() {
     let args = Cli::parse();
     match args.command {
